Create challenge1.sol
This is my answer to the first challenge of Metacrafters in the solidity course.  The challenge stipulates thar I should  write a simple contract and declare four different types of variables. Write get and set functions for each of these variables. Return the value of the variable in the “set function”.

Add your contract to GitHub and share with our Discord channel.